optimization/public-defibrillator-placement/unzipped/requirements.txt
gurobipy==11.0.3
pandas==2.2.2
PyYAML==6.0.1
scipy==1.14.1